requirements-dev.txt
asn1crypto==0.24.0
boto==2.48.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
Django==2.0.5
django-admin-sortable2==0.6.19
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-debug-toolbar==1.11
django-extensions==2.0.7
django-filebrowser-no-grappelli==3.7.4
django-jet==1.0.7
django-model-utils==3.1.2
django-rest-multiple-models==2.1.0
django-simple-history==2.1.1
django-tinymce4-lite==1.7.1
djangorestframework==3.8.2
docutils==0.14
googlemaps==2.5.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
ipython-genutils==0.2.0
Jinja2==2.10
jmespath==0.9.3
jsmin==2.2.2
jsonschema==2.6.0
jupyter-core==4.4.0
nbformat==4.4.0
oauthlib==2.0.7
OWSLib==0.16.0
plotly==2.5.1
pycparser==2.18
pyenchant==2.0.0
Pygments==2.2.0
PyJWT==1.4.0
PyPDF2==1.26.0
python-dateutil==2.7.2
python-decouple==3.1
pytz==2018.4
pyxero==0.9.0
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.3.0
stripe==1.82.1
traitlets==4.3.2
urllib3==1.22
virtualenv==16.0.0
webencodings==0.5.1
xhtml2pdf==0.2.2
